move movieId from moviesSlice to movieDetailsSlice
