+++ src/features/movieDetailsSlice.js
@@ -3,10 +3,14 @@ import { createSlice } from "@reduxjs/toolkit";
 const movieDetailsSlice = createSlice({
     name: 'movieDetails',
     initialState: {
+        movieId: "",
         status: undefined,
         movieDetailsData: {},
     },
     reducers: {
+        setMovieId: (state, { payload: id }) => {
+            state.movieId = id;
+        },
         fetchMovieDetails: state => {
             state.status = "loading";
         },
@@ -21,12 +25,14 @@ const movieDetailsSlice = createSlice({
 });
 
 export const {
+    setMovieId,
     fetchMovieDetails,
     setMovieDetails,
     fetchMovieDetailsError,
 } = movieDetailsSlice.actions;
 
 export const selectMovieDetailsState = state => state.movieDetails;
+export const selectMovieId = state => selectMovieDetailsState(state).movieId;
 export const selectMovieDetailsStatus = state => selectMovieDetailsState(state).status;
 export const selectMovieDetailsData = state => selectMovieDetailsState(state).movieDetailsData;
